Combine deploy and maintain dialogs
